refactor(Board, ChessGame, Main): Board 클래스의 로직 부분을 ChessGame으로 분리
- Board : 8 x 8 보드를 구성하며 체스판을 조작할 수 있는 API 역할
- ChessGame : 체스 게임의 로직을 담당
