+++ src/main/java/softeer2nd/Main.java
@@ -1,6 +1,7 @@
 package softeer2nd;
 
 import softeer2nd.chess.Board;
+import softeer2nd.chess.ChessGame;
 import softeer2nd.chess.ChessView;
 import softeer2nd.chess.Position;
 
@@ -15,6 +16,7 @@ public class Main {
     public static void main(String[] args) {
         String input;
         Board board = new Board();
+        ChessGame chessGame = new ChessGame(board);
         ChessView chessView = new ChessView(board);
         Scanner scanner = new Scanner(System.in);
 
@@ -22,12 +24,12 @@ public class Main {
             while (true) {
                 input = scanner.nextLine();
                 if (input.equals(START_COMMAND)) {
-                    board.initialize();
+                    chessGame.initialize();
                 } else if (input.startsWith(MOVE_COMMAND)) {
                     String[] inputParts = input.split(" ");
                     Position source = Position.create(inputParts[1]);
                     Position destination = Position.create(inputParts[2]);
-                    board.move(source, destination);
+                    chessGame.move(source, destination);
                 } else if (input.equals(END_COMMAND)) {
                     break;
                 }
